requirements.txt
cryptography
pynput
requests
pyautogui
opencv-python
python-dotenv
